chore: move the customize function to the GenericContainerRequest
It will allow providing access to the underlying Logger, the container
Provider, the Started and Reuse states
