+++ generic.go
@@ -5,6 +5,12 @@ import (
 	"errors"
 	"fmt"
 	"sync"
+	"time"
+
+	"github.com/docker/docker/api/types/container"
+	"github.com/docker/docker/api/types/network"
+	"github.com/imdario/mergo"
+	"github.com/testcontainers/testcontainers-go/wait"
 )
 
 var (
@@ -19,6 +25,61 @@ type GenericContainerRequest struct {
 	ProviderType     ProviderType // which provider to use, Docker if empty
 	Logger           Logging      // provide a container specific Logging - use default global logger if empty
 	Reuse            bool         // reuse an existing container if it exists or create a new one. a container name mustn't be empty
+}
+
+// CustomizeRequestOption is a type that can be used to configure the Testcontainers container request.
+// The passed request will be merged with the default one.
+type CustomizeRequestOption func(req *GenericContainerRequest)
+
+// CustomizeRequest returns a function that can be used to merge the passed container request with the one that is used by the container.
+// Slices and Maps will be appended.
+func CustomizeRequest(src GenericContainerRequest) CustomizeRequestOption {
+	return func(req *GenericContainerRequest) {
+		if err := mergo.Merge(req, &src, mergo.WithOverride, mergo.WithAppendSlice); err != nil {
+			fmt.Printf("error merging container request, keeping the original one. Error: %v", err)
+			return
+		}
+	}
+}
+
+// WithImage sets the image for a container
+func WithImage(image string) CustomizeRequestOption {
+	return func(req *GenericContainerRequest) {
+		req.Image = image
+	}
+}
+
+// WithConfigModifier allows to override the default container config
+func WithConfigModifier(modifier func(config *container.Config)) CustomizeRequestOption {
+	return func(req *GenericContainerRequest) {
+		req.ConfigModifier = modifier
+	}
+}
+
+// WithEndpointSettingsModifier allows to override the default endpoint settings
+func WithEndpointSettingsModifier(modifier func(settings map[string]*network.EndpointSettings)) CustomizeRequestOption {
+	return func(req *GenericContainerRequest) {
+		req.EnpointSettingsModifier = modifier
+	}
+}
+
+// WithHostConfigModifier allows to override the default host config
+func WithHostConfigModifier(modifier func(hostConfig *container.HostConfig)) CustomizeRequestOption {
+	return func(req *GenericContainerRequest) {
+		req.HostConfigModifier = modifier
+	}
+}
+
+// WithWaitStrategy sets the wait strategy for a container, using 60 seconds as deadline
+func WithWaitStrategy(strategies ...wait.Strategy) CustomizeRequestOption {
+	return WithWaitStrategyAndDeadline(60*time.Second, strategies...)
+}
+
+// WithWaitStrategyAndDeadline sets the wait strategy for a container, including deadline
+func WithWaitStrategyAndDeadline(deadline time.Duration, strategies ...wait.Strategy) CustomizeRequestOption {
+	return func(req *GenericContainerRequest) {
+		req.WaitingFor = wait.ForAll(strategies...).WithDeadline(deadline)
+	}
 }
 
 // GenericNetworkRequest represents parameters to a generic network
