+++ modules/postgres/postgres.go
@@ -48,7 +48,7 @@ func (c *PostgresContainer) ConnectionString(ctx context.Context, args ...string
 // It will also set the "config_file" parameter to the path of the config file
 // as a command line argument to the container
 func WithConfigFile(cfg string) testcontainers.CustomizeRequestOption {
-	return func(req *testcontainers.ContainerRequest) {
+	return func(req *testcontainers.GenericContainerRequest) {
 		cfgFile := testcontainers.ContainerFile{
 			HostFilePath:      cfg,
 			ContainerFilePath: "/etc/postgresql.conf",
@@ -65,14 +65,14 @@ func WithConfigFile(cfg string) testcontainers.CustomizeRequestOption {
 // It can be used to define a different name for the default database that is created when the image is first started.
 // If it is not specified, then the value of WithUser will be used.
 func WithDatabase(dbName string) testcontainers.CustomizeRequestOption {
-	return func(req *testcontainers.ContainerRequest) {
+	return func(req *testcontainers.GenericContainerRequest) {
 		req.Env["POSTGRES_DB"] = dbName
 	}
 }
 
 // WithInitScripts sets the init scripts to be run when the container starts
 func WithInitScripts(scripts ...string) testcontainers.CustomizeRequestOption {
-	return func(req *testcontainers.ContainerRequest) {
+	return func(req *testcontainers.GenericContainerRequest) {
 		initScripts := []testcontainers.ContainerFile{}
 		for _, script := range scripts {
 			cf := testcontainers.ContainerFile{
@@ -89,8 +89,8 @@ func WithInitScripts(scripts ...string) testcontainers.CustomizeRequestOption {
 // WithPassword sets the initial password of the user to be created when the container starts
 // It is required for you to use the PostgreSQL image. It must not be empty or undefined.
 // This environment variable sets the superuser password for PostgreSQL.
-func WithPassword(password string) func(req *testcontainers.ContainerRequest) {
-	return func(req *testcontainers.ContainerRequest) {
+func WithPassword(password string) testcontainers.CustomizeRequestOption {
+	return func(req *testcontainers.GenericContainerRequest) {
 		req.Env["POSTGRES_PASSWORD"] = password
 	}
 }
@@ -99,8 +99,8 @@ func WithPassword(password string) func(req *testcontainers.ContainerRequest) {
 // It is used in conjunction with WithPassword to set a user and its password.
 // It will create the specified user with superuser power and a database with the same name.
 // If it is not specified, then the default user of postgres will be used.
-func WithUsername(user string) func(req *testcontainers.ContainerRequest) {
-	return func(req *testcontainers.ContainerRequest) {
+func WithUsername(user string) testcontainers.CustomizeRequestOption {
+	return func(req *testcontainers.GenericContainerRequest) {
 		if user == "" {
 			user = defaultUser
 		}
@@ -122,14 +122,16 @@ func RunContainer(ctx context.Context, opts ...testcontainers.CustomizeRequestOp
 		Cmd:          []string{"postgres", "-c", "fsync=off"},
 	}
 
-	for _, opt := range opts {
-		opt(&req)
-	}
-
-	container, err := testcontainers.GenericContainer(ctx, testcontainers.GenericContainerRequest{
+	genericContainerReq := testcontainers.GenericContainerRequest{
 		ContainerRequest: req,
 		Started:          true,
-	})
+	}
+
+	for _, opt := range opts {
+		opt(&genericContainerReq)
+	}
+
+	container, err := testcontainers.GenericContainer(ctx, genericContainerReq)
 	if err != nil {
 		return nil, err
 	}
