+++ modules/redis/redis_test.go
@@ -196,8 +196,10 @@ func TestWithConfigFile(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			req := &testcontainers.ContainerRequest{
-				Cmd: tt.cmds,
+			req := &testcontainers.GenericContainerRequest{
+				ContainerRequest: testcontainers.ContainerRequest{
+					Cmd: tt.cmds,
+				},
 			}
 
 			WithConfigFile("redis.conf")(req)
@@ -232,8 +234,10 @@ func TestWithLogLevel(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			req := &testcontainers.ContainerRequest{
-				Cmd: tt.cmds,
+			req := &testcontainers.GenericContainerRequest{
+				ContainerRequest: testcontainers.ContainerRequest{
+					Cmd: tt.cmds,
+				},
 			}
 
 			WithLogLevel(LogLevelDebug)(req)
@@ -283,8 +287,10 @@ func TestWithSnapshotting(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			req := &testcontainers.ContainerRequest{
-				Cmd: tt.cmds,
+			req := &testcontainers.GenericContainerRequest{
+				ContainerRequest: testcontainers.ContainerRequest{
+					Cmd: tt.cmds,
+				},
 			}
 
 			WithSnapshotting(tt.seconds, tt.changedKeys)(req)
