+++ modules/vault/vault.go
@@ -35,14 +35,16 @@ func RunContainer(ctx context.Context, opts ...testcontainers.CustomizeRequestOp
 		},
 	}
 
-	for _, opt := range opts {
-		opt(&req)
-	}
-
-	container, err := testcontainers.GenericContainer(ctx, testcontainers.GenericContainerRequest{
+	genericContainerReq := testcontainers.GenericContainerRequest{
 		ContainerRequest: req,
 		Started:          true,
-	})
+	}
+
+	for _, opt := range opts {
+		opt(&genericContainerReq)
+	}
+
+	container, err := testcontainers.GenericContainer(ctx, genericContainerReq)
 	if err != nil {
 		return nil, err
 	}
@@ -52,7 +54,7 @@ func RunContainer(ctx context.Context, opts ...testcontainers.CustomizeRequestOp
 
 // WithToken is a container option function that sets the root token for the Vault
 func WithToken(token string) testcontainers.CustomizeRequestOption {
-	return func(req *testcontainers.ContainerRequest) {
+	return func(req *testcontainers.GenericContainerRequest) {
 		req.Env["VAULT_DEV_ROOT_TOKEN_ID"] = token
 		req.Env["VAULT_TOKEN"] = token
 	}
@@ -60,7 +62,7 @@ func WithToken(token string) testcontainers.CustomizeRequestOption {
 
 // WithInitCommand is an option function that adds a set of initialization commands to the Vault's configuration
 func WithInitCommand(commands ...string) testcontainers.CustomizeRequestOption {
-	return func(req *testcontainers.ContainerRequest) {
+	return func(req *testcontainers.GenericContainerRequest) {
 		commandsList := make([]string, 0, len(commands))
 		for _, command := range commands {
 			commandsList = append(commandsList, "vault "+command)
